+++ src/app/[locale]/(lounge)/gallery/page.tsx
@@ -24,12 +24,12 @@ export async function generateMetadata({ params }: Props): Promise<Metadata> {
             url: '/gallery/',
             title: ogTitle,
             description,
-            images: ['/og-cover.png'],
+            images: ['/og-cafelua-entrance-v019.png'],
         },
         twitter: {
             title: ogTitle,
             description,
-            images: ['/og-cover.png'],
+            images: ['/og-cafelua-entrance-v019.png'],
         },
     };
 }